added a new module named 'docmaps'
