+++ data_hub_api/main.py
@@ -2,9 +2,10 @@ import logging
 
 from fastapi import APIRouter, FastAPI, HTTPException
 from fastapi.responses import HTMLResponse
+from data_hub_api.docmaps.codecs.docmaps import ADDITIONAL_MANUSCRIPT_IDS
 
 from data_hub_api.utils.cache import InMemorySingleObjectCache
-from data_hub_api.docmaps.provider import ADDITIONAL_MANUSCRIPT_IDS, DocmapsProvider
+from data_hub_api.docmaps.provider import DocmapsProvider
 
 
 LOGGER = logging.getLogger(__name__)
